requirements/local.txt
-r base.txt

Werkzeug[watchdog]==2.2.3 # https://github.com/pallets/werkzeug
ipdb==0.13.13  # https://github.com/gotcha/ipdb
psycopg2==2.9.5  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
mypy==1.1.1  # https://github.com/python/mypy
django-stubs==1.15.0  # https://github.com/typeddjango/django-stubs
pytest==7.2.2  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.6  # https://github.com/Frozenball/pytest-sugar

# Documentation
# ------------------------------------------------------------------------------
sphinx==6.1.3  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2021.3.14 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8==6.0.0  # https://github.com/PyCQA/flake8
flake8-isort==6.0.0  # https://github.com/gforcada/flake8-isort
coverage==7.2.1  # https://github.com/nedbat/coveragepy
black==23.1.0  # https://github.com/psf/black
pylint-django==2.5.3  # https://github.com/PyCQA/pylint-django
pre-commit==3.1.1  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.2.1  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==3.8.1  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.2.1  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==3.0.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.5.2  # https://github.com/pytest-dev/pytest-django

appnope==0.1.3
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
debugpy==1.6.3
decorator==5.1.1
#Django==4.1.4
djangorestframework==3.14.0
entrypoints==0.4
executing==1.2.0
humanfriendly==10.0
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
jupyter_client==7.4.7
jupyter_core==5.0.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
#packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
#pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.1
tornado==6.2
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
